+++ src/main/java/com/mathfusion/s3/S3Controller.java
@@ -1,6 +1,5 @@
-package com.mathfusion.controller;
+package com.mathfusion.s3;
 
-import com.mathfusion.service.S3Service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
