+++ src/main/java/com/mathfusion/s3/S3Service.java
@@ -1,4 +1,4 @@
-package com.mathfusion.service;
+package com.mathfusion.s3;
 
 import com.amazonaws.HttpMethod;
 import com.amazonaws.services.s3.AmazonS3;
